+++ decidim-budgets/spec/permissions/decidim/budgets/permissions_spec.rb
@@ -32,7 +32,7 @@ describe Decidim::Budgets::Permissions do
     it_behaves_like "permission is not set"
   end
 
-  context "when subject is not a project" do
+  context "when subject is not a project nor an order" do
     let(:action) do
       { scope: :public, action: :vote, subject: :foo }
     end
@@ -48,11 +48,59 @@ describe Decidim::Budgets::Permissions do
     it_behaves_like "permission is not set"
   end
 
-  context "when reporting a budget" do
+  context "when reporting a project" do
     let(:action) do
       { scope: :public, action: :report, subject: :project }
     end
 
     it { is_expected.to eq true }
+  end
+
+  context "when voting a project" do
+    let(:action) do
+      { scope: :public, action: :vote, subject: :project }
+    end
+
+    it { is_expected.to eq true }
+  end
+
+  context "when subject is an order" do
+    let(:action) do
+      { scope: :public, action: :create, subject: :order }
+    end
+
+    it { is_expected.to eq true }
+  end
+
+  context "when component has a parent component" do
+    let(:parent_component) { create :budgets_group_component, :with_children }
+    let(:budget_component) { parent_component.children.first }
+    let(:context) do
+      {
+        current_component: budget_component,
+        project: project,
+        parent_component_context: {
+          workflow_instance: double.tap do |mock|
+                               allow(mock).to receive(:vote_allowed?).and_return(false)
+                             end
+        }
+      }
+    end
+
+    context "when voting a project" do
+      let(:action) do
+        { scope: :public, action: :vote, subject: :project }
+      end
+
+      it { is_expected.to eq false }
+    end
+
+    context "when creating an order" do
+      let(:action) do
+        { scope: :public, action: :create, subject: :order }
+      end
+
+      it { is_expected.to eq false }
+    end
   end
 end