+++ decidim-core/app/models/decidim/component.rb
@@ -9,10 +9,17 @@ module Decidim
     include Publicable
     include Traceable
     include Loggable
+    include ScopableComponent
+
+    has_many :children, foreign_key: "parent_id", class_name: "Decidim::Component", inverse_of: :parent, dependent: :destroy
+    belongs_to :parent, foreign_key: "parent_id", class_name: "Decidim::Component", inverse_of: :children, optional: true
+    delegate :allow_parent?, :allow_children?, to: :manifest
+
+    scope :top_level, -> { where parent: nil }
 
     belongs_to :participatory_space, polymorphic: true
 
-    default_scope { order(arel_table[:weight].asc, arel_table[:manifest_name].asc) }
+    default_scope { order(arel_table[:weight].asc, arel_table[:manifest_name].asc, arel_table[:name].asc) }
 
     delegate :organization, :categories, to: :participatory_space